fix some bugs in invocation service
This pr fixes problems related to notSentInvocations, which happens when
a node is shutdown and an invocation was not sent. When server returns
an error the client will use handleError method instead of returning the
error directly to user. Unused invokeOnTarget method is removed, it will
be added back when necessary.