fix bug in get possible members
Iterating over a map creates a local variable. Taking the address of this local
variable will always point to the same address in memory. Therefore creating a
slice over iterating a map this way was creating a slice with all elements equal
to the last element of map. This commit changes the way slices are merged to fix
this issue.