Return same proxy struct instance from Get calls (#612)
Fixes the following problem. When getting data structures from
client, it was creating a new struct for each call after wrapping
the existing proxy of it. It's a problem for certain data
structures, like PNCounter where it caused having different
vector clocks for PN counters with the same name.

Also includes the following:
* Move linter checks before tests in GH actions definition