Fix members view update listener and handler
Members view update handler incorrectly dealt with MemberInfo
pointers while iterating over the received members view. This
led to corruption of the stored members list.

Another problem was that connection manager was registering
cluster view listeners only on the seed connections while it's
enough to register it only on a single connection. It also
didn't re-register the listener when connections were removed
which led to infinite re-connection attempts.