Introduces Eventually and Never with hazelcast defaults (#622)
Started replacing sleep in a couple of tests with Eventually
and Never

* TestClusterReconnection_ShutdownCluster

* Removing TestClusterReconnection_RemoveMembersOneByOne

This test is not deterministic itself. A client can connect
one of the members while we are shutting them down which
will add another pair of connect/disconnect.
Since we already have TestClusterReconnection_ShutdownCluster
I am removing this test.