Fix Failover (#1028)
The reboot is triggered with the ClusterDisconnected event, which in
turn published when there are no active connections to the cluster.
ClusterDisconnected events may pile up, and cause multiple reboots.
With this change, the number of active connections is checked before
executing the full reboot.

Also, `net.DialTimeout` which is used during the connection to a member
can throw `context.DeadlineExceeded` which causes the circuitbreaker
give up retrying. In this PR, I replace `context.DeadlineExceeded`
returned from `net.DialTimeout` with a custom error.